My Final Efforts
I think the game now has everything we need to demo it and turn it in.
It has 3 levels (they suck, but they are there), almost all of the game
objects we had documented, etc.

If you guys want to try to update it some more go ahead, but I doubt I
will touch it again before it is due.